// src/api/generated/models/GetPaymentsBasic.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { PaymentInList } from './PaymentInList';

export type GetPaymentsBasic = {
  result?: Array<PaymentInList>;
  count: number;
  errors?: Record<string, any>;
};